CIF-1398: Move <AddressForm> component to its own folder from checkout folder
